[Structure] Re-organized structure of pages in project
